refactor: split card-icon into one component per file
